Move files out of individual folders- start file structure refactor
