+++ app/src/components/NavBar.jsx
