+++ app/src/components/index.js
@@ -1,5 +1,5 @@
-export App from './App/App'
-export NotFoundPage from './NotFoundPage/NotFoundPage'
-export NavBar from './NavBar/NavBar'
-export Landing from './LandingPage/LandingComponent'
-export Itinerary from './ItineraryPage/Itinerary.jsx'
+export App from './App.jsx'
+export NotFoundPage from './NotFoundPage.jsx'
+export NavBar from './NavBar.jsx'
+export Landing from './LandingComponent.jsx'
+export Itinerary from './Itinerary.jsx'
