+++ app/src/containers/AppContainer.jsx
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
-import { checkLogin } from '../../actions/auth'
-import { App } from '../../components'
+import { checkLogin } from '../actions/auth'
+import { App } from '../components'
 
 const mapDispatchToProps = (dispatch) => {
   return {
