+++ app/src/containers/index.js
@@ -1,4 +1,4 @@
-export AppContainer from './AppContainer/AppContainer'
-export NavBarContainer from './NavBar/NavBarContainer'
-export LandingContainer from './LandingContainer/LandingContainer'
-export HomePage from './HomePage/HomePage'
+export AppContainer from './AppContainer.jsx'
+export NavBarContainer from './NavBarContainer.jsx'
+export LandingContainer from './LandingContainer.jsx'
+export HomePage from './HomePage.jsx'
